workaround to special character posting
I could not replicate the character and name exactly in my jenkins to match customer, but I pulled in and changed the code into the extension from the underlying library and fixed the incorrect content type setting that I am pretty sure was the problem. I did replicate a problem similar with just a ?, so hopefully testing in customer environment will prove this works.